Move components and rename folders
