+++ src/pages/NewWallet/CheckWordsIntroPage.tsx
