+++ src/pages/NewWallet/CheckWordsPage.tsx
