+++ src/pages/NewWallet/WalletWelcomePage.tsx
