+++ src/pages/NewWallet/WalletWordsPage.tsx
