+++ src/pages/UnlockedWallet/AddressDetailsPage.tsx
