+++ src/pages/UnlockedWallet/OverviewPage/AddressSummaryCard.tsx
@@ -23,11 +23,11 @@ import styled, { css } from 'styled-components'
 
 import { Address } from '@/contexts/addresses'
 
-import AddressBadge from './AddressBadge'
-import Amount from './Amount'
-import ClipboardButton from './Buttons/ClipboardButton'
-import QRCodeButton from './Buttons/QRCodeButton'
-import InputArea from './Inputs/InputArea'
+import AddressBadge from '../../../components/AddressBadge'
+import Amount from '../../../components/Amount'
+import ClipboardButton from '../../../components/Buttons/ClipboardButton'
+import QRCodeButton from '../../../components/Buttons/QRCodeButton'
+import InputArea from '../../../components/Inputs/InputArea'
 
 interface AddressSummaryCardProps {
   address: Address
