+++ src/routes/ImportWalletRoutes.tsx
@@ -20,9 +20,9 @@ import AppHeader from '@/components/AppHeader'
 import FloatingLogo from '@/components/FloatingLogo'
 import { StepsContextProvider } from '@/contexts/steps'
 import { WalletContextProvider } from '@/contexts/wallet'
-import CreateWalletPage from '@/pages/WalletManagement/CreateWalletPage'
-import ImportWordsPage from '@/pages/WalletManagement/ImportWordsPage'
-import WalletWelcomePage from '@/pages/WalletManagement/WalletWelcomePage'
+import CreateWalletPage from '@/pages/NewWallet/CreateWalletPage'
+import ImportWordsPage from '@/pages/NewWallet/ImportWordsPage'
+import WalletWelcomePage from '@/pages/NewWallet/WalletWelcomePage'
 
 const ImportWalletRoutes = () => {
   const importWalletSteps: JSX.Element[] = [
